Backend/requirements.txt
flask
flask-restful
pyjwt
pymysql
PyMuPDF
pytesseract
Pillow
flask_cors
cryptography